requirements.txt
Flask==3.0.3
pandas==2.2.2
matplotlib==3.9.2
mplfinance==0.12.10b0
requests==2.32.3
t-tech-investments
